https://github.com/zurachu/zurapce/pull/6 の内容を反映
